refactor: テストデータのクラスを分離。
